requirements.txt
# Flask Web Framework
Flask==2.3.3
flask-cors==4.0.0

# Neo4j Database Driver
neo4j==5.12.0

# Data Validation
marshmallow==3.20.1

# Environment Variables
python-dotenv==1.0.0

# HTTP Client (for testing)
requests==2.31.0

# Date/Time Utilities
python-dateutil==2.8.2

# Logging (included in Python standard library, but explicit for clarity)
# logging - built-in

# Development Dependencies (optional)
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0